Move login and logout from Header to Sidebar
