+++ src/components/layout/Header.tsx
@@ -1,11 +1,9 @@
 import { AppBar, Box, IconButton, Toolbar } from "@mui/material";
-import LoginIcon from '@mui/icons-material/Login';
 import DehazeIcon from '@mui/icons-material/Dehaze';
 import {useLocation, useNavigate} from "react-router-dom";
 import Tooltip from "@mui/material/Tooltip";
 import {useTranslation} from "react-i18next";
-import {useAuth} from "../../shared/hooks.ts";
-import {AccountCircle, ArrowBack} from "@mui/icons-material";
+import {ArrowBack} from "@mui/icons-material";
 import React from "react";
 
 type Props = {
@@ -15,7 +13,6 @@ type Props = {
 
 const Header = React.forwardRef((props:Props, ref) => {
 
-  const {user} = useAuth(); //Control authentication
   const navigate = useNavigate()
   const location = useLocation();
   const {t} = useTranslation();
@@ -37,42 +34,15 @@ const Header = React.forwardRef((props:Props, ref) => {
             </Tooltip>}
           </Box>
           <Box display={"flex"} sx= {{justifyContent: "right"}}>
-            <Box>
-              {
-                !user ? (
-                  <Tooltip title={t('Sign in.Sign in')}>
-                    <IconButton
-                      size="large"
-                      sx= {{color: "text.secondary"}}
-                      onClick={() => navigate('/signin')}
-                    >
-                      <LoginIcon />
-                    </IconButton>
-                  </Tooltip>
-                ) : (
-                  <Tooltip title={t('Dashboard.YourEvents')}>
-                    <IconButton
-                      size="large"
-                      sx= {{color: "text.secondary"}}
-                      onClick={() => { navigate('/dashboard') }}
-                    >
-                      <AccountCircle />
-                    </IconButton>
-                  </Tooltip>
-                )
-              }
-            </Box>
-            <Box>
-              <Tooltip title={t('Menu')}>
-                <IconButton
-                  size="large"
-                  sx= {{color: "text.secondary"}}
-                  onClick={() => props.setOpenSidebar(prev => !prev)}
-                >
-                  <DehazeIcon/>
-                </IconButton>
-              </Tooltip>
-            </Box>
+            <Tooltip title={t('Menu')}>
+              <IconButton
+                size="large"
+                sx= {{color: "text.secondary"}}
+                onClick={() => props.setOpenSidebar(prev => !prev)}
+              >
+                <DehazeIcon/>
+              </IconButton>
+            </Tooltip>
           </Box>
         </Toolbar>
       </AppBar>
